+++ package.json
@@ -41,9 +41,12 @@
     "query-string": "^7.1.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-icons": "^4.4.0",
     "react-loading": "^2.0.3",
+    "react-loading-skeleton": "^3.1.0",
     "react-redux": "^8.0.1",
     "react-router-dom": "5.3.1",
+    "react-sidebar": "^3.0.2",
     "react-transition-group": "^4.4.5",
     "react-vertical-timeline-component": "^3.5.2",
     "redux": "^4.2.0",
