+++ src/components/Skeleton.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import Skeleton from 'react-loading-skeleton';
+import 'react-loading-skeleton/dist/skeleton.css';
+
+export default function (props) {
+  return <Skeleton {...props} />;
+}
