+++ src/components/index.js
@@ -4,3 +4,4 @@ export { default as AppLoading } from './AppLoading';
 // export { default as CardStack } from './CardStack';
 export { default as Empty } from './Empty';
 export { default as Card } from './Card';
+export { default as Skeleton } from './Skeleton';
