+++ src/layouts/components/DrawerContent.jsx
@@ -6,38 +6,36 @@ import { useQueryParam, NumberParam, StringParam } from 'use-query-params';
 
 import _ from 'lodash';
 
+// atomize
 import {
-  Toolbar,
-  List,
-  ListItem,
-  ListItemButton,
-  ListItemIcon,
-  ListItemText,
-  Divider,
+  Div, Text, Button, Icon, Dropdown, Anchor,
+} from 'atomize';
+
+// custom components
+import {
   Skeleton,
-  FormControl,
-  InputLabel,
-  TextField,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  Button,
-  Select,
-  MenuItem,
-  IconButton,
-} from '@mui/material';
+} from '../../components';
 
-import {
-  BookmarksSharp,
-  AccountCircle,
-} from '@mui/icons-material';
-import FiberManualRecordIcon from '@mui/icons-material/FiberManualRecord';
-import AddIcon from '@mui/icons-material/Add';
-import EditIcon from '@mui/icons-material/Edit';
-import DeleteIcon from '@mui/icons-material/Delete';
+// 菜單項目
+const menuItemStyle = {
+  w: '100%',
+  bg: 'gray100',
+  textColor: 'black700',
+  rounded: '0',
+  d: 'flex',
+  justify: 'flex-start',
+  transition: true,
 
-export default function DrawerContent(props) {
+  hoverBg: 'gray500',
+};
+
+// 菜單項目(on focus)
+const focusMenuItemStyle = {
+  ...menuItemStyle,
+  bg: 'gray400',
+};
+
+function DrawerContent(props) {
   const {
     drawerTypes,
     drawerCategories,
@@ -52,108 +50,51 @@ export default function DrawerContent(props) {
   const [categoryId, setCategoryId] = useQueryParam('CategoryId', NumberParam);
 
   return (
-    <div>
-      <Toolbar />
-      <Divider />
-      <List>
-        <ListItem disablePadding>
-          <ListItemButton
-            to="/notes"
-            component={Link}
-            selected={pathname === '/notes'}
-          >
-            <ListItemIcon>
-              <BookmarksSharp />
-            </ListItemIcon>
-            <ListItemText primary="全部類型" />
-          </ListItemButton>
-        </ListItem>
-        {
-          _.isArray(drawerTypes)
-            ? drawerTypes.map((typeItem, index) => (
-              <ListItem key={typeItem.value} disablePadding>
-                <ListItemButton
-                  to={`/notes/${typeItem.value}${categoryId ? `?CategoryId=${categoryId}` : ''}`}
-                  component={Link}
-                  selected={pathname === `/notes/${typeItem.value}`}
-                >
-                  <ListItemIcon>
-                    <BookmarksSharp />
-                  </ListItemIcon>
-                  <ListItemText primary={typeItem.name} />
-                </ListItemButton>
-              </ListItem>
-            ))
-            : (
-              <>
-                <Skeleton animation="wave" variant="ListItem" />
-                <Skeleton animation="wave" variant="ListItem" />
-                <Skeleton animation="wave" variant="ListItem" />
-              </>
-            )
-        }
-      </List>
-      <Divider />
-      <List>
-        <ListItem disablePadding>
-          <ListItemButton
-            selected={!categoryId}
-            onClick={() => (categoryId ? setCategoryId(undefined) : null)}
-          >
-            <ListItemIcon>
-              <FiberManualRecordIcon />
-            </ListItemIcon>
-            <ListItemText
-              primary="全部分類(自定義分類)"
-            />
-          </ListItemButton>
-        </ListItem>
-        {
-          _.isArray(drawerCategories)
-            ? (
-              <>
-                {
-                  drawerCategories.map((category, index) => (
-                    <ListItem key={category.id} disablePadding>
-                      <ListItemButton
-                        selected={categoryId === category.id}
-                        // eslint-disable-next-line max-len
-                        onClick={() => (categoryId !== category.id ? setCategoryId(category.id) : null)}
-                      >
-                        <ListItemIcon>
-                          <FiberManualRecordIcon style={{ color: category.color }} />
-                        </ListItemIcon>
-                        <ListItemText primary={category.name} />
-                      </ListItemButton>
-                      <IconButton aria-label="edit" onClick={() => openForm({ rowId: category.id })}>
-                        <EditIcon />
-                      </IconButton>
-                      <IconButton aria-label="delete" onClick={() => deleteCategory(category.id, category.name)}>
-                        <DeleteIcon />
-                      </IconButton>
-                    </ListItem>
-                  ))
-                }
-                <ListItem disablePadding>
-                  <ListItemButton
-                    sx={{ display: 'flex', justifyContent: 'center' }}
-                    onClick={() => openForm({})}
-                  >
-                    <ListItemText primary="新增分類" />
-                    <AddIcon />
-                  </ListItemButton>
-                </ListItem>
-              </>
-            )
-            : (
-              <>
-                <Skeleton animation="wave" variant="ListItem" sx={{ marginBottom: 1, padding: 3 }} />
-                <Skeleton animation="wave" variant="ListItem" sx={{ marginBottom: 1, padding: 3 }} />
-                <Skeleton animation="wave" variant="ListItem" sx={{ marginBottom: 1, padding: 3 }} />
-              </>
-            )
-        }
-      </List>
-    </div>
+    <Div
+      bg="gray100"
+      h="100%"
+    >
+      <Button
+        {...pathname === '/notes' ? focusMenuItemStyle : menuItemStyle}
+      >
+        全部類型
+      </Button>
+      {
+        _.isArray(drawerTypes)
+          ? drawerTypes.map((typeItem, index) => (
+            <Button
+              key={typeItem.value}
+              {...pathname === `/notes/${typeItem.value}` ? focusMenuItemStyle : menuItemStyle}
+            >
+              {typeItem.name}
+            </Button>
+          ))
+          : (
+            <Skeleton count={3} />
+          )
+      }
+      <hr />
+      <Button
+        {...(!categoryId && pathname.includes('/notes')) ? focusMenuItemStyle : menuItemStyle}
+      >
+        全部分類(自定義分類)
+      </Button>
+      {
+         _.isArray(drawerCategories)
+           ? drawerCategories.map((category, index) => (
+             <Button
+               key={category.id}
+               {...(categoryId === category.id && pathname.includes('/notes')) ? focusMenuItemStyle : menuItemStyle}
+             >
+               {category.name}
+             </Button>
+           ))
+           : (
+             <Skeleton count={3} />
+           )
+      }
+    </Div>
   );
 }
+
+export default DrawerContent;
