+++ src/layouts/components/index.js
@@ -1,2 +1,4 @@
 export { default as DrawerContent } from './DrawerContent';
-export { default as CategoryFormModal } from './CategoryFormModal';
+// export { default as CategoryFormModal } from './CategoryFormModal';
+export { default as AppSidebar } from './AppSidebar';
+export { default as AppBar } from './AppBar';
